@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.0.2",
-    "@angular/cdk": "^17.0.0",
-    "@angular/common": "17.0.2",
-    "@angular/compiler": "17.0.2",
-    "@angular/core": "17.0.2",
-    "@angular/forms": "17.0.2",
-    "@angular/material": "^17.0.0",
-    "@angular/material-moment-adapter": "17.0.0",
-    "@angular/platform-browser": "17.0.2",
-    "@angular/platform-browser-dynamic": "17.0.2",
-    "@angular/router": "17.0.2",
+    "@angular/animations": "18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "18.0.0",
+    "@angular/compiler": "18.0.0",
+    "@angular/core": "18.0.0",
+    "@angular/forms": "18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/material-moment-adapter": "18.0.0",
+    "@angular/platform-browser": "18.0.0",
+    "@angular/platform-browser-dynamic": "18.0.0",
+    "@angular/router": "18.0.0",
     "core-js": "^2.4.1",
     "cors": "^2.8.5",
     "express": "^4.16.2",
@@ -34,11 +34,11 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.0",
-    "@angular-devkit/schematics": "^17.0.0",
-    "@angular/cli": "^17.0.0",
-    "@angular/compiler-cli": "17.0.2",
-    "@angular/language-service": "17.0.2",
+    "@angular-devkit/build-angular": "^18.0.1",
+    "@angular-devkit/schematics": "^18.0.1",
+    "@angular/cli": "^18.0.1",
+    "@angular/compiler-cli": "18.0.0",
+    "@angular/language-service": "18.0.0",
     "@types/express": "^4.0.39",
     "@types/jasmine": "~3.8.0",
     "@types/jasminewd2": "~2.0.2",
@@ -52,6 +52,6 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ts-node": "~3.2.0",
-    "typescript": "5.2.2"
+    "typescript": "5.4.5"
   }
 }
