@@ -25,7 +25,7 @@ import { MatSortModule } from "@angular/material/sort";
 import { MatTableModule } from "@angular/material/table";
 import { MatToolbarModule } from "@angular/material/toolbar";
 import {CoursesService} from "./services/courses.service";
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { CourseDialogComponent } from './course-dialog/course-dialog.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {LoginComponent} from './login/login.component';
@@ -44,8 +44,7 @@ import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {OnlyOneErrorPipe} from './pipes/only-one-error.pipe';
 import {LoginReactiveComponent} from './login-reactive/login-reactive.component';
 
-@NgModule({
-    declarations: [
+@NgModule({ declarations: [
         AppComponent,
         HomeComponent,
         AboutComponent,
@@ -63,10 +62,8 @@ import {LoginReactiveComponent} from './login-reactive/login-reactive.component'
         LoginReactiveComponent,
         OnlyOneErrorPipe
     ],
-    imports: [
-        BrowserModule,
+    bootstrap: [AppComponent], imports: [BrowserModule,
         BrowserAnimationsModule,
-        HttpClientModule,
         MatMenuModule,
         MatButtonModule,
         MatIconModule,
@@ -90,12 +87,9 @@ import {LoginReactiveComponent} from './login-reactive/login-reactive.component'
         MatStepperModule,
         MatProgressBarModule,
         FormsModule,
-        ReactiveFormsModule
-    ],
-    providers: [
-        CoursesService
-    ],
-    bootstrap: [AppComponent]
-})
+        ReactiveFormsModule], providers: [
+        CoursesService,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule {
 }
