@@ -1,7 +1,7 @@
 
 
 import {Injectable} from "@angular/core";
-import {HttpClient, HttpParams} from "@angular/common/http";
+import { HttpClient, HttpParams } from "@angular/common/http";
 import {Observable} from "rxjs";
 import {Course} from "../model/course";
 import {map} from "rxjs/operators";
